+++ jetbrains-toolbox/src/preload.js
@@ -1,28 +1,32 @@
 const fs = require("fs");
-const { exec } = require("child_process");
-const softwares = require("./softwares")
+const { execFile } = require("child_process");
+const softwares = require("./softwares");
 
-utools.onPluginReady(() => {
-});
-
+utools.onPluginReady(() => {});
 
 function getHistory() {
-  let allHistory = []
+  let allHistory = [];
   // [{title description icon exec }]
-  let toolboxHistorys = JSON.parse(fs.readFileSync(utools.getPath("home") + "/Library/Caches/JetBrains/Toolbox/intellij_projects.json").toString());
-  fs.writeFile
+  let toolboxHistorys = JSON.parse(
+    fs
+      .readFileSync(
+        utools.getPath("home") +
+          "/Library/Caches/JetBrains/Toolbox/intellij_projects.json"
+      )
+      .toString()
+  );
+  fs.writeFile;
   toolboxHistorys.forEach((toolboxHistory, index) => {
-    let history = {}
-    history.title = toolboxHistory.name
-    history.description = toolboxHistory.path
+    let history = {};
+    history.title = toolboxHistory.name;
+    history.description = toolboxHistory.path;
 
-    let software = softwares[toolboxHistory.defaultOpenItem.application_id]
-    history.icon = software.icon
-    history.exec = software.exec
-    allHistory.push(history)
-  })
-  console.log(allHistory)
-  return allHistory
+    let software = softwares[toolboxHistory.defaultOpenItem.application_id];
+    history.icon = software.icon;
+    history.exec = software.exec;
+    allHistory.push(history);
+  });
+  return allHistory;
 }
 
 let History = {
@@ -35,17 +39,23 @@ let History = {
 
     search: (action, searchWord, callbackSetList) => {
       if (!searchWord) return callbackSetList(getHistory());
-      return callbackSetList(getHistory().filter((x) => x.title.toLowerCase().indexOf(searchWord.toLowerCase()) !== -1
-        || x.description.toLowerCase().indexOf(searchWord.toLowerCase()) !== -1));
+      return callbackSetList(
+        getHistory().filter(
+          (x) =>
+            x.title.toLowerCase().indexOf(searchWord.toLowerCase()) !== -1 ||
+            x.description.toLowerCase().indexOf(searchWord.toLowerCase()) !== -1
+        )
+      );
     },
 
     select: (action, itemData) => {
-
-      let command = `"${utools.getPath("home") + itemData.exec}" "${itemData.description}"`;
-      console.log(command, 123)
-      exec(command, (err) => {
-        if (err) utools.showNotification("不是有效的可执行程序");
-      }); // 这种命令必须异步执行
+      execFile(
+        utools.getPath("home") + itemData.exec,
+        [itemData.description],
+        (err) => {
+          if (err) utools.showNotification("不是有效的可执行程序");
+        }
+      ); // 这种命令必须异步执行
 
       utools.outPlugin();
       utools.hideMainWindow();
@@ -54,5 +64,5 @@ let History = {
 };
 
 window.exports = {
-  History
+  History,
 };