+++ jetbrains-toolbox/CHANGELOG.md
@@ -1,5 +1,8 @@
 # 更新日志
 
-## v0.0.1（当前版本）
+## v1.0.0（当前版本）
 
-使用默认配置打开历史项目
+1. 仅支持MAC平台
+2. 仅支持通过JetBrains Toolbox安装的IDE打开的项目
+3. 仅支持默认的安装路径
+4. 目前支持IDE：Idea、Goland、DataGrip、WebStorm